requirements.txt
Flask
mysql-connector-python
python-dotenv
google-generativeai
streamlit
requests